requirements.txt
Django
gunicorn
python-decouple
psycopg[binary]
dj-database-url
Pillow
whitenoise
django-cloudinary-storage
requests